refactor: Move consul scripts to custom consul img
The configuration files that were located in the root build directory
that were referenced by a separate running consul image have been
removed. They were added to the new consul image that is a part of this
issue (issue-220). The Dockerfile was also updated to remove all
mentions of consul - as it no longer is relevant.

Signed-off-by: Beau Frusetta <beau.frusetta@intel.com>